Ok. Here are the ACTUAL changes... forgot to add the files
No longer create the application and db at the beginning of the file. This cleans up at lot of the code and the structure
as there will be much less circular imports and creating new database modules should be easy. Instead, database creation
is done in extensions.py and imported. When application.py is called the application is created and returned, there are now
a number of configuration functions in init so that the application can be create with different configurations.
The purpose of this was so that we can create a separate application with different configs (different db file etc...) for
testing purposes. This also cleans up a lot of the logic around __init__.py

Different configurations can be inserted into config.py, which can then be imported and used with the application.

Worked with Oseme to make the above changes.