Merged the followers module with the user module
The followers module was basically just a table defining a relationship between users so I thought it made more sense to merge them. Added a 'follow'/'unfollow' beside courses and the code behind it so that users can follow each other. This does not yet affect the order in which reviews are displayed. Also created a few unit tests for followers.