Review sorting by follower and vote rating also a lot of refactoring
Reviews should now be displayed first in order of followers and next in order of vote rating. I refactored a lot of the follower code... took a painful amount of time but I think it is better this way. The main point is that most of the sorting now happens server side and not client side. Changed things to user sendObjects() instead of getJSON to facilitate tests to come later.