(iOS) Added login view and now able to test if login passed or failed to the local host using json.
Changed the login function on the server side so it can send a json object to the iOS so it knows if login passed or failed (Not sure if there was another way of doing this without changing the server side). Still need to test if the session is stored and if post and follow functions work.